#4 -Reorganizando estrutura de autenticacao por token
