+++ package.json
@@ -7,7 +7,7 @@
     "test": "mocha --timeout 10000 --recursive --reporter nyan",
     "testd": "mocha --debug-brk --timeout 10000 --recursive --reporter nyan",
     "coveralls": "istanbul cover ./node_modules/mocha/bin/_mocha --report lcovonly -- -R spec && cat ./coverage/lcov.info | ./node_modules/coveralls/bin/coveralls.js && rm -rf ./coverage",
-    "start": "node src --debug-brk",
+    "start": "node src",
     "debug": "DEBUG=* node src"
   },
   "repository": {
